Fix user profile bug
The user profile consists of the posts made by the user whose profile is being visited (apart from other details).
Now, the posts to be displayed in the profile page should be visible to both -
the authenticated user and the user whose profile is being visited
i.e. it should consist of posts made by the user whose profile is being visited in the classrooms where both of them are a part of, either
as students or teacher-student. This was not case earlier. This commit fixes that bug.